font-patcher: Remove option --use-single-width-glyphs
[why]
There are three possible options to specify the Nerd Font Mono generation:
    -s
    --mono
    --use-single-width-glyphs

All the three are handled the same.

In order to add a new long option that also handles glyph width in a "single"
cell manner the fear is that two too similar options will confuse users.

[how]
Just hide the longest form.
Also remove from the readme files.

Need to adapt the install.sh script.

(The option still works but is not 'advertised'.)

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>