Devicons: Manually optimize many icons
[why]
A lot icons have visually obvious issues, small but visible.
Often the source is a color svg and someone just made them monochrome
for the plain version. But it is still different paths and not joined.
That can lead to small glitches with the fontforge svg import.

[how]
Manually inspect some icons and recreate a new plain icon wich consists
of one flat path.

For Plotly we also update to the current logo with circles as ...
circles and not some squiggly lines.

Tomcat and Zig also get plain variants which they lack upstream.

Fixes: #1771

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>