Devicons: Correct version number [skip-ci]
[why]
The version number of the deficons.ttf did not change but the glyphs got
modified.

[how]
The version number is manual anyhow, so use a `git describe --tags` like
descriptor. That can also stay for example when/after 3.4.0 is released.

[note]
Skipping zip creation because another MR will be pulled soon and we can
reduce the repo size by having fewer changed files (i.e. the zip).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>