doc: Add hint for Homebrew on Linux
[why]
If people want to use Homebrew on Linux to install the fonts they fail
following the documentation.

The reason is that Casks are not normally considered on Linux, because
they contain binary blobs for MacOS by definition. Well, the fonts are
indeed binaries, but could still be installed on any platform.

I can not find any expressive statement in the Homebrew or Homebrew-Cask
documentation, but trying on MacOS 15.3 and Ubuntu 24.04 showed this
different behavior.

[how]
Add back the `--cask` but only as addition for Linux, because
 - Main users of Homebrew are MacOS, and there it is deprecated
 - It might change on the Linux platform as well (for fonts?)

[note]
PR #1789
PR #1079

Reported-by: Pablo Sanchez @pablogrs
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>