Switch octicons file to otf
[why]
We construct the octicons font file from the original svgs.
To put them into the font file the svg is interpreted and transformed
into the font coordinate system, and then again scaled.

The resulting font file is then saved as `ttf`, thereby rounding all
coordinates. That font is then used as source for the real patching
process - and will be rounded again (possibly).

The rounding errors accumulate and we should only round once. That means
intermediate fonts need to be `otf`.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>