font-patcher: Create smaller otf files
[why]
The patched fonts in otf format are much bigger than the ttf
counterparts. Even when the same font is used as source.
The factor is about 2 to 3.5 in size.

[how]
Fonts in ttf format can not store nodes with sub-integer precision. The
coordinates are rounded on export. But otf fonts can have real numbers
(may it be floating point or fixed point notation, I am not sure right
now and it does not matter) for coordinates, but the more precise
numbers take much more space in the font file. If an otf is created
where the coordinates are already integer this extra space is not
needed.

So we do all the calculations with real numbers (including decimals) and
only in the last step the coordinates of the added glyphs are rounded to
integer (i.e. to the EM size grid).

Fixes: #1848

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>