font-patcher: Fix rounded scales in borderline cases
[why]
The unrounded scaling is more or less exact. But if we round the
coordinates it can happen that the leftmost node moves to the left and
the rightmode node moves to the right (due to the rounding), so that the
width in total grew by more than 1 EM-unit. This is escpecially bad for
the Nerd Font Mono fonts where the whole glyph should be within the
bounding box.

[how]
Check the bounding box after scaling and retry (paste again and scale
again) with a one unit smaller aim in the rare cases that the problem
occurs.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>