font-patcher: Reconsider borderline scale issues
The previous solution is maybe too complicated and again assumes to
much. For example it allows glyphs to be up to (but less than) one
em-unit too wide. Because we warn only then, but even if it is 0.8
em-units wider than intended it is still too wide.

So instead we aim (calculate) to a width that is one em-unit smaller.
This means under all circumstances the resulting glyph will not wider
than our target (even with rounding problems), but some will be a tiny
bit smaller than possible.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>