font-patcher: Improve node-rounding on Mono fonts [skip ci]
[why]
Sometimes, due to rounding issues, we get monospaced glyphs that have a
small negative rhs bearing (i.e. extend out of the cell). That should
not be the case and all of the glyph should be inside the 'cell' (i.e.
advance width).

[how]
After changing the glyph size to the final size (before shifting) the
actual bounding box is checked, and when it extends outside the expected
limits the glyph is rescaled (from the original glyphs, not scaled a
second time) a bit smaller. As much smaller as the first try was too
big.
That can again fail due to rounding, so we try this three times with
increasing magnitude of the correction.

[note]
Also change the Monospace glyph size check for Mono fonts, so that
glyphs with (positive) overlap are never checked.
This is skipped because we do not care if the overlap is one em-unit
more or less, we are too big already anyhow. And re-scaling takes time
that can be saved.

This fixes PR #1851

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>