install.sh: Reintruduce extension parameter
[why]
At the moment we have every font only in one format (ttf or otf), so the
preference option does not do much (does not do anything).

Anyhow, it was there before and maybe at some point we take the multiple
formats up again, and then we are prepared.

[how]
Instead of the previous solution to search for all formats and then try
to remove the duplicates we search first for the desired format and if
we find not one font we search for the other format. This is slightly
different but the outcome is the same.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>