install.sh: Make bash 3 compatible
[why]
Vanilla MacOS has just some very old bash 3 installed by default. People
might want to use the install.sh, but not install a newer bash (via
Homebrew for example).

[how]
The `nerdfonts_dirs` array is replaced by a normal bash variable (string)
where the directories are '/' separated. Because there are never
subdirectories in the list but only direct children of 'patched-fonts/'
it is possible to use slash as unique and out-of-band separator.

The `fonts` array is replaced by a function that delivers the list of
files in null-byte separated form. That means we can not manipulate the
list and thus the duplicate checking is dropped. As we do not have any
font in both formats (anymore) it will never happen anyhow.
Dropping the duplicates check also removes the extension preference
option functionality. We keep the now defunct option, though.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>