CI: Create new FontPatcher.zip only at the end of release workflow
[why]
We have an obsolete glyhnames.json in the release FontPatcher.zip.

[how]
Packaging the FontPatcher.zip should be the last step when everything
has already been updated.

Some code from the zip-release.yml workflow is copied into the
release.yml :-/ but we need to do it at the exact right time.

Note that we need to commit-push before we call the
github-pages-deploy-action.

[note]
Related: #1868

Reported-by: @subframe7536
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>