font-patcher: Fix wrong specialities matching
[why]
When the name to parse has 'for' in it - even when it is in the middle
of a word like 'formidable', 'unfortunate', or, 'forced' the "For" part
will be cut off and separated from the rest.

[how]
The reason is the code to remove 'For Powerline' from font names.

Instead of matching 'For' and 'Powerline' individually and then removing
the 'Powerline' and 'For' token later on (which fails if there is just a
'For' token like with 'Un' 'For' 'Tunate'), we now match only on a 'For'
if it is followed by a 'Powerline'.

To make the removal of the other-token easier later on we rename/unify
the token to be CamelCase.

Fixes: #1877

Reported-by: @mietzen
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>